Add portrait layout and allow system to set orientation
When using a phone, landscape mode is not always desirable.
This update leaves orientation up to the system rather than
forcing landscape.                            
                                                  
It also adds a portrait layout which places the buttons below
the word entry and increases the width of the word list. The
original landscape mode is moved to layout-land, but otherwise
untouched.